Include SMS costs map in dynamic landscape facility
The facility to change the input SMS costs map is provided in batch mode only as part of the dynamic landscape feature; the costs may be changed simultaneously with other landscape changes or independently of them. For the GUI, there remains no option to change the cost map during the course of a simulation.